Remove broken component classes.
Fixes #12. Non-existant templates were used as layouts.
As the classes themselves were empty and since ember-modals-manager-internal
already exports the "parent" component classes they can be removed completely.

The issue described in #12 did not occur in the dummy app and in tests
as they were not using the component classes from this addon, but the
ones from ember-modals-manager-internal going under the same name.